requirements.txt
cycler==0.10.0
kiwisolver==1.3.1
matplotlib==3.3.4
numpy==1.20.0
Pillow==8.1.0
pyparsing==2.4.7
python-dateutil==2.8.1
scipy==1.6.0
six==1.15.0
Wave==0.0.2
firebase_admin
requests
python-firebase